+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-30 15:22:22
-// Commit: chore(security): update dependencies and cleanup unused imports in security
+// Updated at: 2026-05-30 21:46:21
+// Commit: style(security): harmonize light and dark theme palette for security
 
 class securityComponentState {
-  final String commitRef = '1624';
-  final String lastMessage = 'chore(security): update dependencies and cleanup unused imports in security';
-  final DateTime timestamp = DateTime.parse('2026-05-30T15:22:22+05:00');
+  final String commitRef = '1646';
+  final String lastMessage = 'style(security): harmonize light and dark theme palette for security';
+  final DateTime timestamp = DateTime.parse('2026-05-30T21:46:21+05:00');
 }